requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pymongo==4.6.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.7
